Undo scaling and translation before running Xfoil or writing coords
These operations could cause a slight change in Cm due to the fact that
Cm is always calculated at x=0.25, z=0. The translation and scaling is
needed for creating the shape functions (which go from 0 to 1), but for
consistency's sake it is best to undo them when passing off the
coordinates to Xfoil.